Refactor CLI tools. Get rid of configuration.

+++ gstandaard/parsing.py
@@ -1,10 +1,10 @@
 import os
 import requests
 from bs4 import BeautifulSoup
+import argparse
 
 from .files import file_numbers
 from .constants import SKIP_FIELDS, BESCHRIJVINGEN_FILE_EXTENSION
-from .config import BESCHRIJVINGEN_DIRECTORY, BESTAND_BESCHRIJVINGEN_URL_PREFIX, DATA_DIRECTORY
 
 
 def get_bestand_filename(name, add_ext=False):
@@ -14,20 +14,25 @@ def get_bestand_filename(name, add_ext=False):
     return filename
 
 
-def get_bestand_htmlpath(name):
-    return os.path.join(BESCHRIJVINGEN_DIRECTORY, get_bestand_filename(name, add_ext=True))
+def get_bestand_htmlpath(directory, name):
+    return os.path.join(directory, get_bestand_filename(name, add_ext=True))
 
 
 def get_bestand_url(file_no):
+    BESTAND_BESCHRIJVINGEN_URL_PREFIX = 'https://www.z-index.nl/documentatie/bestandsbeschrijvingen/bestand?bestandsnaam='
     return BESTAND_BESCHRIJVINGEN_URL_PREFIX + get_bestand_filename(file_no)
 
 
-def get_inputpath(file_no):
-    return os.path.join(DATA_DIRECTORY, get_bestand_filename(file_no, add_ext=False))
+def get_inputpath(directory, file_no):
+    return os.path.join(directory, get_bestand_filename(file_no, add_ext=False))
 
 
 # https://techoverflow.net/2017/02/26/requests-download-file-if-it-doesnt-exist/
-def download_file(filename, url):
+def download_file(filename, url, refresh):
+
+    if os.path.exists(filename) and not refresh:
+        return
+
     with open(filename, 'wb') as fout:
         response = requests.get(url, stream=True)
         response.raise_for_status()
@@ -35,26 +40,25 @@ def download_file(filename, url):
             fout.write(block)
 
 
-def download_if_not_exists(filename, url):
-    if not os.path.exists(filename):
-        download_file(filename, url)
-        return True
-    return False
+#
+# Console script entry point
+#
+def download_htmls():
+    ap = argparse.ArgumentParser()
+    ap.add_argument("-b", "--beschrijvingen-directory", required=True,
+                    help="Directory with HTML beschrijvingen files")
+    ap.add_argument('-r', '--refresh', action='store_true',
+                    help='Download files even if they exist already.')
+    args = vars(ap.parse_args())
 
-
-import click
-@click.command()
-@click.option('--refresh', is_flag=True, default=False, help='Download files even if they exist already.')
-def download_htmls(refresh):
     for file_no in file_numbers:
-        if refresh:
-            download_file(get_bestand_htmlpath(file_no), get_bestand_url(file_no))
-        else:
-            download_if_not_exists(get_bestand_htmlpath(file_no), get_bestand_url(file_no))
+        download_file(get_bestand_htmlpath(args['beschrijvingen_directory'], file_no),
+                      get_bestand_url(file_no),
+                      args['refresh'])
 
 
-def get_bestand_html(file_no):
-    html_input = get_bestand_htmlpath(file_no)
+def get_bestand_html(directory, file_no):
+    html_input = get_bestand_htmlpath(directory, file_no)
     html = BeautifulSoup(open(html_input).read(), features='html5lib')
 
     return html
